Set up the initial framework for the game
This sets up the basic framework we need in order to start on the actual
logic of doing something with the game.

The idea that we're going with here is that the game will be contained
in its own window, which we control by hiding all of the chrome and
blocking some key commands.

The result is a plain window that the user can't generally interact with
(for the most part) except for key bindings that we will eventually
create.