Fix _is_sudoku for windows with no active view
I'm not 100% sure why this is possible, but I noticed that I was getting
errors out of this function call because events were triggering for
windows that didn't have an active view.

This fixes that to report such windows as not being sudoku views,
although I'm not entirely sure why that's necessary.