Implement Sudoku grid creation
After playing around with a few designs, this one seems the most like
what we want for now. Ultimately I would like to use line drawing
characters for this, but there is currently a bug in the build of
Sublime that I'm using that makes that not a good idea.