Include a debug flag to re-enable commands
As one might suspect, being unable to select text, view scopes or run
commands is somewhat problematic when it comes to debugging things.

For this reason, we now don't consider a view to be a sudoku view if it
has a debug setting enabled. This allows us to enable that setting for
the view so that we can regain control.