Augment the syntax for better game display
This adds a few simple rules into the syntax so that we can syntax
highlight on the grid and the values stored within it.

With the simple rule set provided here, we can detect when digits are
being used as hints in the grid and display them as comments, while for
sequences of two or more of the same digit, we flag them with a color
that indicates that the cell has been filled.