Rework the characters in the grid
In order to be able to easily discern all of the cells, it would be nice
if there was a regex pattern that would match only at the top left of
the cells. The original grid layout did not allow for this because it
was ambiguous whether the sequence `+-` was in the top or bottom of a
cell (and so the last line of a set of three in a group would get
matched by the rule).

This alters the display a little bit so that this is no longer ambiguous
and things will work the way we want. When the bug with box drawing
characters is resolved, we can easily swap the correct characters in
for higher visual fidelity.