Include an event handler to close the window
This includes some new event handlers for close actions so that if the
user closes the Sudoku view, the whole window will close if it is now
empty of views of any kind.

This makes quitting the game a little nicer if you press the wrong key
sequence because the dedicated Sudoku window has no chrome.