Render the grid with the new functionality
Now that we have a mechanism for getting text into the view, this shifts
the code for displaying the grid so that it's a call to the render
method instead of an inline call to append.

This requires a slight modification to the render code because it wants
to gather the list of cells on first call, which it can't do it if it
hasn't written the cells yet.