Include a new base command for refactoring
Our current code structure has us dispatching methods in the rendering
class based on an argument provided using a simple lookup mechanism, and
for clarity and extensibility this seems to work rather well.

Although we could use a single command for all of the logic and
rendering, it seems like it would make more logical sense to have the
game logic in one command and the rendering in another.

This commit adds in a new base class that those two classes can derive
from so that we don't have to duplicate code.

This currently wraps the existing logic for how this works in the
current render command, as well as some additional helpers for
determining regions based on cell offsets that may come in handy.