Add in a command to drive the game logic
Using our new base, this implements a new command that will drive the
game logic based on user actions. 

This command will be responsible for holding the board and all of the
game state, and will drive the render command in order to realize state
changes.

For the moment the only game state that is stored is the puzzle that is
being played, and a notion of a "current" grid position where future
user input will affect the board.

The currently available actions here are starting a new game in the view
and allowing the input position to change.