Refactor the render command
Continuing on in our refactoring of the existing code, this modifies the
render code to use the new base, which removes some redundant code from
the class and the global headers.

In addition to that the puzzle rendering action has been modified
because under the new scheme the puzzle state is stored in the game
command, so the state to render needs to be provided as an argument to
the action instead.

A new action to use regions to highlight the currently active grid
position has also been added.