Make the game state persistent
In the previous test implementation, the state of the game was stored in
instance variables in the command, which is good enough for testing but
not a particularly robust solution otherwise because over restarts or
when the plugin reloads, the state is lost.

This change stores the game state into view settings so that they will
persist. We also set the regions that we're using to mark the selection
as being persistent as well so that on restart they will still be
visible.