Factor away the edit objects
A `TextCommand` receives an `Edit` object when it is invoked, and that
object is needed to modify the buffer. That object is only valid for the
current command invocation, and is used by Sublime to group edits in the
undo stack.

In the current design, it's only needed in the render command because
only that command is going to edit text, but not all of the render
actions need it, and it's certainly not needed for the game logic either
but it needs to be passed around because both classes conform to a
common base.

This small change factors the edit object out into an instance variable
that is updated on invocation so that code that needs it can have it
but it doesn't need to be passed around, which makes everything a lot
cleaner looking.