Alter rendering to use a square font
I wasn't happy with the visualization of the grid being wider than it
is tall just to make the grid cells look more square, and since this is
meant to be a game and not a standard plugin I feel that it should be
OK to specify a custom font.

To that end I traced down a square font meant for roguelikes so we're
using that now. A downside to this is that we can't use box drawing
characters because the font doesn't support them, and the fallback font
is not square so things don't line up properly.

As such we're going to leave the current visualization as is, although
we have fixed the grid size as a result.