Include the ability to render with hinting
The idea of the controls of the game are going to be that arrow keys
will move the selection cursor around, and pressing the number keys will
either add/remove a hint or add/remove an answer from a cell.

To that end we now track a "hinting" state, and this state tracks what
color is used to generate the cell highlight.