Include input support
Two new bindings are now active to allow us to capture user input while
the game is focused. Both go through the main logic command but with
slightly different actions so that we can separate the input for the
normal mode from the input for the hinting mode.

Both modes respond to a space and invoke a new action that toggles the
state of the highlight, which is a nice test that the input mechanism is
working.