Modify rendering to look a little nicer
This change encompasses a series of rendering changes that bring us
closer to what the final version is going to look like.

Mainly this alters the text generated into the cells so that we can
visually distinguish right answers, wrong answers and hints that appear
in the middle of the cell with different colors.

In order to pull this off we also need to pack in a color scheme
override to allow us to style some extra scopes. Currently this is being
done by forcing the color scheme to be Monokai and having an override,
but a more final version would pack in a small color scheme instead or
create one based on the values in the user's current color scheme.

For the moment the answers are being marked as alternatively correct and
incorrect in order to test the styling.