Add state and fix a rendering bug
First off, there was a slight issue with the existing render code in
which only the initial render could modify the buffer but future renders
could not. This is because the view setup makes the view `readonly`.
Weirdly, the commands for rendering could modify the buffer in its
initial state but not once the first render call was made.

This defect was fixed by making the render method temporarily flip the
view state.

In addition, the game now tracks the state of guesses and hints and the
rendering has been improved in order to support display of that
information.

Lastly, for debugging purposes we've added in a redraw command which
redraws the grid (and is the reason we found that bug above) so that
while tweaking things we can see what the view looks like.