Allow entering hints and answers
This implements the logic which allows the user to enter a hint for a
cell or an answer, as well as removing an answer. The logic for
determining the correctness of the answer is still missing however. In 
its place for debugging purposes we're allowing a shifted version of the
number to set an incorrect state.

When the input is a 0, the answer is removed and any hints that used to
be at that location are re-displayed; there is a binding that binds this
to the backspace key since that seems more natural.