Properly validate the board after a move
This implements a board validation routine that checks all of the cells
to see if they're invalid or not and then flags the state as appropriate.

This works by checking each row, column and 3x3 grid segment and
checking to see what values are invalid for each of those items because
they appear more than once.

Once that is done, we have three lists, each indexable by 0 through 8
that represent the three states, and in each list is the values of
invalid items.

All that we have to do is check each board position to see if the value
there appears in any of the three lists, and if so it's invalid.