Clean up a hinky inline of
I'm fairly certain that lack of sleep and pondering on the problem of
the clearest way to evaluate a Sudoku board to find errors may have
caused me a bit of mental distress, which is carrying on into today.

After struggling for a bit too long last night and tonight clarity
finally hit when I realized how simple this whole thing could actually
be.