Include some stubbed out resource files
This includes a settings file as well as integrations with the main
menu and command palette for the GZipper package. This is currently a
little more than we need, since at the moment I'm not sure that we need
to have key bindings at all, but we have the menu entries for it here so
we can remove them later as needed.

This also includes a stub README file, but since we're not 100% sure yet
how all of this is going to finally come together, it's pretty sparse at
the moment.