Implement a setting to control auto-unzipping
This adds a simple setting which can be configured to determine if the
package should automatically try to `gunzip` every file that opens, if
it looks like the file is gzipped.

When set to `true` (which is the default), this is enabled and so every
file opened will be checked to see if it is a gzipped file.

When set to `false`, the automatic functionality does not trigger, which
requires that the user manually `gunzip` the file instead by choosing
the appropriate entry from the command palette.