Provide more feedback to the user
This provides some extra feedback to the user when working with a
gzipped file.

A file that represents an unzipped version of a file is tagged as such
in the status bar so that it's easy to see at a glance what's going on,
and when saving a file we also display a status message that mentions
that we're recompressing the file.