Include a command to create a compressed file
This ties all of the existing code together by making it possible for
you to create a new gzipped file from the file that you're currently
editing.

This can be done in two ways; the first is to just create a one-off
compressed version of the file. In this case the original file remains.
The other way is to convert the current view to a gzip as if it was a
gzip all along.

In this case the file will be recompressed on every save and deleted
when it's closed, leaving only the gzipped version behind.