Fix silly bug introduced during some refactoring
While I was cleaning up code last night in preparation for the final
series of commits, I accidentally optimized away a variable name that
was actually needed for the plugin to work.

This puts it back (storing the name of the created archive) so that we
can indicate that we're compressing.