All for resaving of a gzipped temporary file
As previously defined, when you open a gzipped file you actually open a
temporary file which is recompressed on save and deleted on close.

If you were to save-as the file from its old name to a new name, then
the result was previously that saving the new file still recompressed to
the original gzipped file, since the name of that file is carried as a
setting.

This change has a view that is a temporary file track the name that it
was opened with, so that an `on_pre_save` event handler can detect that
the name has changed and break the connection between the current file
and the underlying file.

This allows you to save-as an uncompressed file into a new file, leaving
the original file alone (which seems to be what one would probably
expect in this situation).