Include some threading logic
Include the basis of a background working thread for our
compress/uncompress operations.

This creates an abstract base class that can be easily subclassed to do
some work. When the thread is created, it immediately launches, and once
the work is complete a callback is invoked in the main Sublime thread to
allow the code to finish up whatever it needs to do with the resulting
data.