Create output files for unzip operations in tmpdir
Using the power of `tempfile.mkstemp()`, this makes the output files
that we generate when we open a compressed file use a temporary file
that is created in the temporary folder, which ensures that we don't
collide with anything that already exists.

The file is created in the temporary directory with a name and extension
that mimics the underlying file, but which has a random suffix (instead
of `_ungz` which we used previously), which makes sure that files are
not accidentally clobbered when we write the output files.