Handle potential trash errors by removing files
Under Linux, the `send2trash` library tries to put things into the Trash
by putting them into a `.Trash-####` folder. The folder is placed at a
logical place on the same mount point as the source file so that it can
do a simple move operation to put the file in the trash.

Normally this works as one might expect, but for cases where a user's
home directory is on a different mount point than the root file system,
(or different from where `/tmp` is mounted), this results in trying to
put the `.Trash` folder in the root of the hard drive, which fails due
to a permission error.

This change implements a new `delete_on_trash_fail` setting which
defaults to `false`. When it's `true`, when we try to trash a file and
it fails, the package will physically delete the file instead (otherwise
it will just leave it in place).

This allows the user to decide if they want to keep files around or not.