Add an additional setting to control trashing
After some consideration, I have decided that perhaps a user may want to
control whether a file gets put to the trash or flat out deleted,
instead of assuming the trash.

In that case, our code needs to delete the file directly instead of
trashing it, or if the trash fails delete it as we did before, based on
the value for a setting.