Fix a small bug in gzip_compress
The command for compressing a current file was enabling or disabling
itself based on the file being a gzip file or not (i.e. don't enable for
something that is already a gzip file).

That's a good idea, but it glosses over the notion that the command gets
the name of the output file by appending an extension to the input file,
so if the current file hasn't been saved yet, that can't happen.

This solves the issue by ensuring that the buffer also already has a
filename.