Verify with the user before clobbering a gzip file
We have a command that lets you promote a text file to a gzipped file or
just do a one-off compression. This command uses the name of the current
file to create the archive. It's possible that the archive file already
exists, in which case this operation would clobber the file.

This modifies that command to check and see if the file that would be
created actually exists or not. If it does, instead of clobbering we are
asked to confirm,and only on confirmation does the operation happen.

Here we've done this by adding an argument to the command to tell it
that it can force the operation through without confirmation, and the
command invokes itself with that as an argument (along with what it was
originally invoked with) in order to carry out the action.