Include a new setting for an upcoming feature
We're about to implement a feature that would allow us to remove a gzip
file from disk, optionally also leaving the temporary file around (which
could be used to drop the gzip file entirely).

Since we may be deleting the temporary file on user confirmation, this
new setting will control whether we force close the temporary view or
leave it open.

By default it will stay open and appear unsaved (because the underlying
file is gone), which could be used to recover. Setting the value to
true will force the buffer closed.