Include a new command to remove the archive file
This command will remove the `gzip` archive file from disk, and
optionally also remove the temporary editing file as well (or depending
on your point of view, optionally leave it alone).

This uses the newly added setting to control what happens to the buffer
if the user chooses to delete the temporary file.

The command has been added to the command palette in two variations, and
it confirms what is going to happen before it carries out the action.