Include sample code from the API library
This includes a slightly modified version of the sample code for the
Google API library, which is handling things like OAUTH and such for us.

As currently written, this will let you log in via a browser and then
display the list of videos into the console, so long as you massage the
script a bit to make it run.

This is in no way ready for prime time, since it runs on plugin load
and hangs Sublime while you log in. It also should be caching the login
token somehow, although I'm not sure at the moment if that's possible.

Based on a lot of research and fiddling around I'm reasonably sure this
also should not be including a client secret at all, since this is going
to be distributed to people and they need the secret to pretend to be
us, but that's what the wizard provided.