Include the client secret and update the path
After some poking, it seems like perhaps this client secret is not as
secret as I initially thought, or at least it seems to know that this is
an installed application (as opposed to a web application) and I assume
that Google knows that any file packed in an application is subject to
someone finding it.

As such, it's now added. In addition the location to the file in the
package is now figured when the plugin loads so that instead of being
hard coded to my local source directory it is sourced from the package
itself.

For the time being the package needs to be installed unpacked since the
Google libraries are loading the file locally.