Include credential caching
This includes some caching of credentials as they are requested, so that
subsequent runs can continue to use them instead of constantly forcing
the user to log in again.

I'm unsure at the moment how this would react if the token that's cached
gets out of date; perhaps this needs to update the cache on every
request to keep the token active, or maybe that's not needed. Time will
tell.

The documentation certainly won't, that much is clear (not that I'm
bitter, mind you).