Include encryption of the cached tokens
It seems like somewhat of a good idea to encrypt the token and refresh
token that we get back when the user authenticates and authorizes our
application, since that token wraps up the ability to ask for access and
get it on our behalf (or at least, that's how I understand it).

For expediency here we're going for using pure Python AES encryption and
a pure Python PBKDF implementation because otherwise we need to try and
compile native code using the same ABI as Sublime is using, and nobody
needs that. The encryption needs we have are pretty light.

For the moment, the password is hard coded just for expediency in our
testing, but in the future it will be something that the user provides
us as needed.