Clean up our cache data a bit
Although the `token` and `refresh_token` are needed, I don't think that
the `id_token` is needed (particularly since we don't get one anyway),
so for the time being we're going to redact it from the cached data.

In addition, we don't need to include `client_secret`, since we're using
the Installed Application flow which doesn't use it anyway (since it's
not secure), so we can drop that too.