Include the start of our own networking thread
This starts the ground work for having our own background thread that
will be doing all of the work of making requests. The ultimate goal is
that the front end commands will communicate with the thread by giving
it instructions on what operations are required, and the thread will
tell them what the results of that request are.