Include a request dispatch mechanism
Include the beginnings of our message request dispatch mechanism. We
allow a call to specify a request (currently a string) and the callback
that should be invoked when that request is completed. 

The request is dropped into a Queue, where the network code will pull it
out and then immediately dispatch a result by using `set_timeout()` to
shift the code into the `main` Sublime thread.