Start moving the authorization code to networking
To begin working on actually have commands execute in the network
thread, this moves the authentication code into the networking code
(or more specifically it copies it there; the original command is still
in place for testing purposes at the moment although it will not work).

This allows the thread to be told to authorize itself using the new
mechanism.