Implement request callback interception
In order for the network manager to be able to keep track of state and
do caching, we first need a mechanism in place for us to catch the
callback being invoked so that we can introspect it and act as
appropriate.

Here we're doing that by changing the callback that gets invoked by the
network thread so that it calls our own callback, which can look at the
data and then invoke the user callback transparently.