Implement the notion of caching of a result
This is far from complete at the moment, but the idea here is that the
manager code has a cache of results, and every time a request is made
and our callback gets called, we capture the value of the return. 

Later calls will immediately get the cached result back instead of
making a new request, unless a refresh is specifically requested by the
caller.

At the moment this is just caching the results on success, which is
perhaps not what we want. Also the whole mechanism needs to be a bit
more general.