Update cached credentials to fix restart problems
After far too long, the source of the issue I was having has been
discovered and ultimately the reason for the weird transient failure
under the new code is that the call to create the cached credentials
was not correct as it was missing a field.

I used to include that field but removed it since it was always None,
but it may be a good idea to include it anyway.

In addition I've discovered that even though the documentation for
OAuth2 seems to indicate that the client secret should not be needed, in
fact if it is not included the refresh token doesn't work.

At the moment I'm unsure of what this means; should I be using some sort
of external web application somewhere even though this is a client side
app, or what? For our purposes here, that's less of an issue, I think.