Update our caching mechanism
What we have now is the notion that whenever we make a request, if it
succeeds we should cache the result, but if we fail we should remove any
cached result that may exist on a previous call.

The logic here is that any request that fails may need to be undertaken
again, in which case we should ask again and not cache the failure.

The request method has been streamlined so that for any request that
exists in the cache, the cached value is returned so long as the refresh
action was not provided (if the user wants to refresh, then reqeust
anew).

With this in place the logic that tries to determine if we need to
actually make the request no longer needs to special case the authorize
request like it used to.