Update the callback to get the request made
This augments the callback mechanism between the network thread/manager
and the client so that when the client callback is invoked, it gets the
request that it made along with the result. This makes it easier for the
client to share a single callback, which allows it to chain requests if
needed.