Include a Request object
Now that we have our mechanism working a little bit better, this
modifies it so that instead of passing a string, we pass a dictionary
instead.

The idea here is that some requests need to provide not only the name of
a request, but extra arguments that indicate how to fetch that data out.

So our actual Request object is a subclass of `dict` that is hashable
to allow this to work. It inherently knows about the name as a property,
and must be constructed with a name and an optional list of keys and
values, which will be added as arguments for the handler to do with as
it likes.