Rewrite our command using our new API
This commit (which is the last for Day 25 despite the time stamp on it)
rewrites the existing list command to use our new API, which gets us off
of the async thread but otherwise performs more or less like the
previous version (sans error checking).

This still requires some cleanup, error handling and spinners, but it
does prove that indeed our concept works.