Clean up the initial command
This cleans up our initial command to make it a little easier to read by
abstracting away all of the common code into a single method for making
a request.

There is likely a better way to do the sort of chained response that we
are doing here, but this is good enough for the time being.