Create some request abstraction
This adds a first pass for a bit of a wrapper around the workings of our
mechanism for making network requests.

The idea here is that any commands that will be making requests will be
doing so using the same general mechanism, so having it abstracted away
seems like the way to go.

Additionally, in order to stop the mess of long ifelse chains, our
Request class can now be told what method should be used to handle it,
and the abstractions will use that to dispatch the result.

What we have currently is the idea that every command will have an
`_error` method to handle errors made during requests, and otherwise
they can set up methods that assume when they are called that the
request has succeeded, which makes the code a little cleaner.