Demand-start the YouTube thread
As a simple change, this alters our model so that the network code
doesn't start up the background thread needed to talk to anything until
the first request is made, so that it's not sitting idle in the
background for no good reason.

This is probably not of huge concern, but as an old-timey kind of guy it
upsets my delicate sensibilities to have idle threads for no reason.