Include a new request to deauthenticate
This adds in a new request to the network thread to request that it
deauthenticate itself. The response to this request is to discard the
cached copy of the login credentials and discard the cached YouTube
service (if any).

Once this action is taken, no further YouTube requests can be made until
the user logs in again, which goes through the steps of getting the user
to authorize that the plugin is allowed to get data from that YouTube
account.