Streamline our command a bit
This is a small simple change and is actually something that was present
during testing but got lost in the various changes being made.

The idea here is that the network manager knows if the user is currently
authenticated or not (which here means that a request could be made and
it would be serviced; so the user has authorized the app previously and
something has already started a request previously).

So, if that is the case we can go directly to asking for the data that
we want, and only have to authorize when it's needed. That saves a
little bit of back-and-forth time on requests.