Move command enabled state to the helper
We're currently working under the assumption that except for the command
that authorizes the plugin to access a YouTube account, any command will
want to only be enabled when when those credentials are available.

As such, our common helper class now implements a version of
is_enabled() for use in all subclasses that want that, and any that want
to alter it can provide it explicitly.