Abstract required authorization away a bit
Most commands that are requesting data need to use the `authorize`
request to ensure that they are authorized, since this starts up the
network thread (if it's not already), gets the user to authorize the
plugin (if that hasn't already happened) and sets up the service in the
thread that will make requests.

Based on that, instead of having every command contain an initial
if/else chain we instead have a common method that returns True when the
user is authorized already and False when not.

When False is returned, authorization happens automatically; thus in the
`run` method you can just say "if authorized, make this request".

This could be a little more streamlined because we need to ensure that
subclasses implement `_authorize` currently; this can be cleaned up once
there are more commands making requests like these and we have a better
sense of what the pattern should be.