Add extra abstraction for data commands
It wasn't sitting right with me that every command would have a `run`
method that should call a method to see if it was authorized, and have
that method be the thing that does the authorization if it's not already
done because this duplicates code.

This changes things so that the helper class contains a `run` method
that implicitly either runs authorization or triggers a predefined
handler. The internal auth version stores the authorization request and
result so that it can be provided to the outer authorization handler.

From a command standpoint, instead of having a `run` method it can just
assume that if `_authorized()` is called, it's authorized to make
requests and can go ahead.